+++ crypto/evp/evp_lib.c
@@ -168,17 +168,18 @@ int EVP_CIPHER_type(const EVP_CIPHER *ctx)
 	}
 }
 
-int EVP_CIPHER_block_size(const EVP_CIPHER *e)
+size_t EVP_CIPHER_block_size(const EVP_CIPHER *e)
 	{
 	return e->block_size;
 	}
 
-int EVP_CIPHER_CTX_block_size(const EVP_CIPHER_CTX *ctx)
+size_t EVP_CIPHER_CTX_block_size(const EVP_CIPHER_CTX *ctx)
 	{
 	return ctx->cipher->block_size;
 	}
 
-int EVP_Cipher(EVP_CIPHER_CTX *ctx, unsigned char *out, const unsigned char *in, unsigned int inl)
+size_t EVP_Cipher(EVP_CIPHER_CTX *ctx, unsigned char *out,
+		  const unsigned char *in, size_t inl)
 	{
 	return ctx->cipher->do_cipher(ctx,out,in,inl);
 	}
@@ -208,22 +209,22 @@ void EVP_CIPHER_CTX_set_app_data(EVP_CIPHER_CTX *ctx, void *data)
 	ctx->app_data = data;
 	}
 
-int EVP_CIPHER_iv_length(const EVP_CIPHER *cipher)
+size_t EVP_CIPHER_iv_length(const EVP_CIPHER *cipher)
 	{
 	return cipher->iv_len;
 	}
 
-int EVP_CIPHER_CTX_iv_length(const EVP_CIPHER_CTX *ctx)
+size_t EVP_CIPHER_CTX_iv_length(const EVP_CIPHER_CTX *ctx)
 	{
 	return ctx->cipher->iv_len;
 	}
 
-int EVP_CIPHER_key_length(const EVP_CIPHER *cipher)
+size_t EVP_CIPHER_key_length(const EVP_CIPHER *cipher)
 	{
 	return cipher->key_len;
 	}
 
-int EVP_CIPHER_CTX_key_length(const EVP_CIPHER_CTX *ctx)
+size_t EVP_CIPHER_CTX_key_length(const EVP_CIPHER_CTX *ctx)
 	{
 	return ctx->key_len;
 	}
@@ -238,7 +239,7 @@ int EVP_CIPHER_CTX_nid(const EVP_CIPHER_CTX *ctx)
 	return ctx->cipher->nid;
 	}
 
-int EVP_MD_block_size(const EVP_MD *md) 
+size_t EVP_MD_block_size(const EVP_MD *md) 
 	{
 	return md->block_size;
 	}
@@ -253,7 +254,7 @@ int EVP_MD_pkey_type(const EVP_MD *md)
 	return md->pkey_type;
 	}
 
-int EVP_MD_size(const EVP_MD *md)
+size_t EVP_MD_size(const EVP_MD *md)
 	{
 	if (!md)
 		return -1;
